SO-2022: Change type of complex map group and map priority to integer
- this is necessary to support values greater than Byte.MAX_VALUE (127)

https://snowowl.atlassian.net/browse/SO-2022

+++ snomed/com.b2international.snowowl.snomed.importer.rf2/src/com/b2international/snowowl/snomed/importer/rf2/UnsupportedRf2RefSetMember.java
@@ -93,12 +93,12 @@ public abstract class UnsupportedRf2RefSetMember extends FakeEObjectImpl impleme
 	}
 
 	@Override
-	public void setMapGroup(byte value) {
+	public void setMapGroup(int value) {
 		throw new UnsupportedOperationException();
 	}
 
 	@Override
-	public void setMapPriority(byte value) {
+	public void setMapPriority(int value) {
 		throw new UnsupportedOperationException();		
 	}
